% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Alaska}
\alias{Alaska}
\title{A boundary file for Alaska, USA}
\format{A \code{SpatialPolygons} object with slots @polygons, @plotOrder, @bbox, and @proj4string. The projection is in Alaska Alber, EPSG:3338.}
\usage{
data(Alaska)
}
\description{
A boundary file for Alaska, USA, of object type \code{SpatialPolygons} from \code{sp} package.
}
\examples{
data(Alaska)
summary(Alaska)
library(sp)
plot(Alaska)
}
\keyword{datasets}
