% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CAKR}
\alias{CAKR}
\title{A boundary file for Cape Krusenstern National Moment, Alaska, USA}
\format{A \code{SpatialPolygonDataFrame} object with slots @polygons, @plotOrder, @bbox, and @proj4string. The projection is in Alaska Albers, EPSG:3338.}
\source{
Obtained from \href{https://catalog.data.gov/dataset/national-parks}{data.gov}. The National Parks dataset September 30, 2019 is part of the U.S. Department of Transportation (USDOT)/Bureau of Transportation Statistics's (BTS's) National Transportation Atlas Database (NTAD). These are the National Park Service unit boundaries. These park boundaries signify legislative boundary definitions and local park names have been consolidated according to the legislation.  The Cape Krusenstern National Monument was subsetted from the all NPS boundaries by importing with the \code{rgdal} package as a \code{SpatialPolygonDataFrame} object from package \code{sp}.
}
\usage{
data(CAKR)
}
\description{
A boundary file for Cape Krusenstern National Moment, Alaska, USA, with object type \code{SpatialPolygonDataFrame} from \code{sp} package.
}
\examples{
data(CAKR)
summary(CAKR)
library(sp)
plot(CAKR)
}
\keyword{datasets}
