% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caribouDF}
\alias{caribouDF}
\title{Data for a caribou forage experiment}
\format{A \code{data.frame} with 30 records and 7 variables:
\describe{
  \item{i}{The ith row if envisioning the grid of plots as a matrix}
  \item{j}{The jth column if envisioning the grid of plots as a matrix}
  \item{water}{A factor with level "N" if no water was added, and a "Y" if water was added.}
  \item{tarp}{A factor with 3 levels: "clear" for a clear tarp over the plot, "none" as a control, and "shade" for a shade tarp over the plot}
  \item{trt}{A factor with 6 levels, treating each unique combination of water and tarp factors as a treatment level}
  \item{z}{The response variable, which was the percent nitrogen in prostrate willows in the 3rd clipping of the second year of the study}
  \item{y}{The y-coordinates of the plots in 2-D space}
  \item{x}{The x-coordinates of the plots in 2-D space}
}}
\source{
These data were provided by Elizabeth Lenart of the Alaska Department of Fish and Game.  The data were used in the publication below.
\cr \cr
The R code that creates the data file from raw data can be found at \code{system.file("raw data/caribouForage/create_caribouDF.R",package = "ZVHdata")}
}
\usage{
data(caribouDF)
}
\description{
Data for a caribou forage experiment in Alaska, where 30 plots were placed in the field in a 5 x 6 grid and a two factor experiment, one with water added or not, and another with 3 different tarp treatments, were applied.
}
\examples{
library(ZVHdata)
data(caribouDF)
names(caribouDF)
summary(caribouDF)
summary(lm(z ~ water + tarp + water:tarp, data = caribouDF))
}
\references{
Lenart, E.A., Bowyer, R.T., Ver Hoef, J.M. and Ruess, R.W. 2002. Climate Change and Caribou: Effects of Summer Weather on Forage. Canadian Journal of Zoology 80: 664-678.
(\href{https://www.nrcresearchpress.com/doi/abs/10.1139/z02-034#.XvweB3VKjmF}{link to: Climate Change and Caribou: Effects of Summer Weather on Forage})
}
\keyword{datasets}
