% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOSSpreds}
\alias{MOSSpreds}
\title{Prediction sites for heavy metals in mosses near a mining road in Alaska, USA}
\format{
A \code{sf} object. The projection is in Alaska Albers, EPSG:3338.  The dataframe has 2357 observations and 5 variables.
\describe{
  \item{lon}{Longitude, in decimal degrees}
  \item{lat}{Latitude, in decimal degrees}
  \item{dist2road}{Distance to haul road, in meters}
  \item{sideroad}{A factor variable with two levels: "N" for north of the haul road, and "S" for south of the haul road}
  \item{strat}{A stratification variable for prediction sites, in distance classes from the haul road, where 1 is the closest, and 5 is the farthest, from the road}
}
}
\source{
This prediction grid was provided by Peter Neitlich and used in the publication given below. 
\cr \cr
Imported to \code{R} with the \code{rgdal} package as a \code{SpatialPoints} object from package \code{sp}. The code that creates the data file from the raw data can be found at \code{system.file("raw data/MOSSHeavyMetals/create_MOSSpreds.R",package = "ZVHdata")}. It was subsequently converted to \code{sf} using \code{st_as_sf} function.
}
\usage{
data(MOSSpreds)
}
\description{
Prediction sites for heavy metal concentrations in mosses near a mining road in Alaska, USA with object type \code{SpatialPointsDataFrame} from \code{sp} package. Data on covariates used to model observed data are included with site location.
}
\examples{
library(ZVHdata)
library(sf)
data(MOSSpreds)
names(MOSSpreds)
summary(MOSSpreds)
data(CAKRboundary)
plot(CAKRboundary)
plot(MOSSpreds[MOSSpreds$strat==5,], add = TRUE, pch = 19, cex = .5, col = 'green')
plot(MOSSpreds[MOSSpreds$strat==2,], add = TRUE, pch = 19, cex = .2, col = 'blue')
plot(MOSSpreds[MOSSpreds$strat==1,], add = TRUE, pch = 19, cex = .2, col = 'red')
}
\references{
Neitlich, P.N., Ver Hoef, J.M., Berryman, S. D., Mines, A., Geiser, L.H., Hasselbach, L.M., and Shiel, A. E. 2017. Trends in Spatial Patterns of Heavy Metal Deposition on National Park Service Lands Along the Red Dog Mine Haul Road, Alaska, 2001-2006. PLOS ONE 12(5):e0177936 DOI:10.1371/journal.pone.0177936
(\href{https://journals.plos.org/plosone/article/file?id=10.1371/journal.pone.0177936&type=printable}{link to: Trends in Spatial Patterns of Heavy Metal Deposition on National Park Service Lands Along the Red Dog Mine Haul Road, Alaska, 2001-2006.})
}
\keyword{datasets}
