% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointSimCSR.R
\name{pointSimCSR}
\alias{pointSimCSR}
\title{Simulate clustered spatial point pattern}
\usage{
pointSimCSR(npair = 100, lower.x.lim = -1, upper.x.lim = 1,
  lower.y.lim = -1, upper.y.lim = 1)
}
\arguments{
\item{npair}{number of points to add that are completely spatially random (CSR), default = 100}

\item{lower.x.lim}{left limit of boundary, default = -1}

\item{upper.x.lim}{right limit of boundary, default = 1}

\item{lower.y.lim}{lower limit of boundary, default = -1}

\item{upper.y.lim}{upper limit of boundary, default = 1}
}
\value{
data.frame of two columns, x-coordinate in the first, and y-coordinate in the second
}
\description{
simulates a clustered spatial point pattern
}
\author{
Jay Ver Hoef
}
