% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulo.R
\name{pulo}
\alias{pulo}
\title{Predict Method for Spatial Linear Mixed Model Fits}
\usage{
pulo(cope_out, pred_data, theta = cope_out$theta,
  extrap = cope_out$extrap, spatial_model = cope_out$spatial_model,
  z = NULL, X = NULL, Xp = NULL, Zs = NULL, Zsp = NULL,
  xcoords = NULL, ycoords = NULL, xcoordsp = NULL, ycoordsp = NULL,
  predmeth = "alldata", nNN = 50, par = FALSE)
}
\arguments{
\item{x}{A fitted splm object}

\item{predData}{A data.frame object to be used for prediction.  It must have the same columns and names for the response variable, any covariates, and x-coordinates and y-coordinates.}
}
\value{
\code{predict.splm} produces a data.frame with 2 columns for predictions, and prediction standard errors for missing response variable, one with the response variable name appended by \code{.pred} for predictions and by \code{.predSE} for the prediction standard errors.
}
\description{
Makes predictions for missing data or using external prediction data set.
}
\details{
These are the universal kriging equations found in Cressie (1993, pg. 154-155).
}
\references{
\cite{Cressie, N.A.C. (1993) Statistics for Spatial Data. Wiley.}
}
\author{
Jay Ver Hoef
}
