% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdobs}
\alias{rdobs}
\title{Data on heavy metals in lichens near a mining road in Alaska, USA}
\format{A \code{SpatialPointsDataFrame} object with slots @data, @coords.nrs, @coords, @bbox, and @proj4string. The projection is in Alaska Albers, EPSG:3338.  The @data is a dataframe with 365 observations and 15 variables.
\describe{
  \item{sample}{A factor variable with location identifier.  Some samples were duplicated in the field, and replicated in the lab, so there are 318 unique spatial locations}
  \item{lon}{Longitude, in decimal degrees}
  \item{lat}{Latitude, in decimal degrees}
  \item{dist2road}{Distance to haul road, in meters}
  \item{sideroad}{A factor variable with two levels: "N" for north of the haul road, and "S" for south of the haul road}
  \item{field_dup}{Integer for any field duplicate samples. All locations start with 1, and any additional samples continue with integers 2, ...}
  \item{lab_rep}{Integer for any replicated samples for laboratory analysis, where the sample was split and a separate laboratory analysis was performed. All analyses start with 1, and any additional analyses continue with integers 2, ...}
  \item{Cd}{Cadmium concentration in moss tissue, in mg/kg}
  \item{Pb}{Lead concentration in moss tissue, in mg/kg}
  \item{Zn}{Zinc concentration in moss tissue, in mg/kg}
}}
\usage{
data(rdobs)
}
\description{
Data on heavy metal concentrations in mosses near a mining road in Alaska, USA with object type \code{SpatialPointsDataFrame} from \code{sp} package. Data on three metals: 1) Cd (cadmium), 2) Pb (lead), and 3) Zn (zinc) are included, along with covariates that have relationships to heavy metal concentrations.
}
\examples{
data(rdobs)
names(rdobs@data)
summary(rdobs)
library(sp)
data(CAKR)
plot(CAKR)
plot(rdobs, add = TRUE, pch = 19)
}
\references{
Neitlich, P.N., Ver Hoef, J.M., Berryman, S. D., Mines, A., Geiser, L.H., Hasselbach, L.M., and Shiel, A. E. 2017. Trends in Spatial Patterns of Heavy Metal Deposition on National Park Service Lands Along the Red Dog Mine Haul Road, Alaska, 2001–2006. PLOS ONE 12(5):e0177936 DOI:10.1371/journal.pone.0177936
(\href{https://journals.plos.org/plosone/article/file?id=10.1371/journal.pone.0177936&type=printable}{link to: Trends in Spatial Patterns of Heavy Metal Deposition on National Park Service Lands Along the Red Dog Mine Haul Road, Alaska, 2001–2006.}) \cr Hasselbach, L., Ver Hoef, J.M., Ford, J., Neitlich, P., Berryman, S., Wolk B. and Bohle, T. 2005. Spatial Patterns of Cadmium, Lead and Zinc Deposition on National Park Service Lands in the Vicinity of Red Dog Mine, Alaska. Science of the Total Environment 348: 211–230.
(\href{https://www.sciencedirect.com/science/article/pii/S0048969705000082}{link to: Spatial Patterns of Cadmium, Lead and Zinc Deposition on National Park Service Lands in the Vicinity of Red Dog Mine, Alaska.})
}
\keyword{datasets}
