% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdpreds}
\alias{rdpreds}
\title{Prediction sites for heavy metals in mosses near a mining road in Alaska, USA}
\format{A \code{SpatialPointsDataFrame} object with slots @data, @coords.nrs, @coords, @bbox, and @proj4string. The projection is in Alaska Albers, EPSG:3338.  The @data is a dataframe with 2357 observations and 4 variables.
\describe{
  \item{lon}{Longitude, in decimal degrees}
  \item{lat}{Latitude, in decimal degrees}
  \item{dist2road}{Distance to haul road, in meters}
  \item{sideroad}{A factor variable with two levels: "N" for north of the haul road, and "S" for south of the haul road}
  \item{strat}{A stratification variable for prediction sites, in distance classes from the haul road, where 1 is the closest, and 5 is the farthest, from the road}
}}
\usage{
data(rdpreds)
}
\description{
Prediction sites for heavy metal concentrations in mosses near a mining road in Alaska, USA with object type \code{SpatialPointsDataFrame} from \code{sp} package. Data on covariates used to model observed data are included with site location.
}
\examples{
data(rdpreds)
names(rdpreds@data)
summary(rdpreds)
library(sp)
data(CAKR)
plot(CAKR)
plot(rdpreds[rdpreds$strat==5,], add = TRUE, pch = 19, cex = .5, col = 'green')
plot(rdpreds[rdpreds$strat==2,], add = TRUE, pch = 19, cex = .2, col = 'blue')
plot(rdpreds[rdpreds$strat==1,], add = TRUE, pch = 19, cex = .2, col = 'red')
}
\references{
Neitlich, P.N., Ver Hoef, J.M., Berryman, S. D., Mines, A., Geiser, L.H., Hasselbach, L.M., and Shiel, A. E. 2017. Trends in Spatial Patterns of Heavy Metal Deposition on National Park Service Lands Along the Red Dog Mine Haul Road, Alaska, 2001–2006. PLOS ONE 12(5):e0177936 DOI:10.1371/journal.pone.0177936
(\href{https://journals.plos.org/plosone/article/file?id=10.1371/journal.pone.0177936&type=printable}{link to: Trends in Spatial Patterns of Heavy Metal Deposition on National Park Service Lands Along the Red Dog Mine Haul Road, Alaska, 2001–2006.}) \cr Hasselbach, L., Ver Hoef, J.M., Ford, J., Neitlich, P., Berryman, S., Wolk B. and Bohle, T. 2005. Spatial Patterns of Cadmium, Lead and Zinc Deposition on National Park Service Lands in the Vicinity of Red Dog Mine, Alaska. Science of the Total Environment 348: 211–230.
(\href{https://www.sciencedirect.com/science/article/pii/S0048969705000082}{link to: Spatial Patterns of Cadmium, Lead and Zinc Deposition on National Park Service Lands in the Vicinity of Red Dog Mine, Alaska.})
}
\keyword{datasets}
