% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SO4pred}
\alias{SO4pred}
\title{Prediction sites for SO4 atmospheric deposition throughout the conterminous USA}
\format{A \code{SpatialPoints} object with slots  @coords, @bbox, and @proj4string. The projection is in Conus (Continental US) Albers, EPSG:5070. There are 2357 locations in the @coords slot for prediction.}
\usage{
data(SO4pred)
}
\description{
Prediction sites for SO4 atmospheric deposition throughout the conterminous USA with object type \code{SpatialPoints} from \code{sp} package.
}
\examples{
data(SO4pred)
str(SO4pred)
summary(SO4pred)
library(sp)
data(USboundary)
plot(USboundary)
plot(SO4pred, add = TRUE, pch = 19, cex = .5)
}
\keyword{datasets}
