% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SO4preds}
\alias{SO4preds}
\title{Prediction sites for SO4 atmospheric deposition throughout the conterminous USA}
\format{
A \code{sf} object. The projection is in Conus (Continental US) Albers, EPSG:5070. There are 9558 locations for prediction.
}
\source{
The code that creates the data file from a systematic grid within the conterminous US can be found at \code{system.file("raw data/wetSulfateDep/create_SO4preds.R",package = "ZVHdata")}. It was subsequently converted to \code{sf} using \code{st_as_sf} function.
}
\usage{
data(SO4preds)
}
\description{
Prediction sites for SO4 atmospheric deposition throughout the conterminous USA with object type \code{sf} from \code{sf} package.
}
\examples{
library(ZVHdata)
library(sf)
data(SO4preds)
str(SO4preds)
summary(SO4preds)
data(USboundary)
plot(st_geometry(USboundary))
plot(st_geometry(SO4preds), add = TRUE, pch = 19, cex = .5)
}
\keyword{datasets}
