% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AKboundary}
\alias{AKboundary}
\title{A boundary file for Alaska, USA}
\format{
A \code{sf} object. The projection is in Alaska Albers, EPSG:3338.
\describe{}
}
\source{
Shapefile downloaded from
\cr \cr
\href{https://www.arcgis.com/home/item.html?id=b07a9393ecbd430795a6f6218443dccc}{link to: Shapefile of 50 US States hosted by ArcGIS}. 
Imported to \code{R} with the \code{rgdal} package as a \code{SpatialPolygonsDataFrame} object from package \code{sp}, subsetted to just the Alaska polygon, and then simplified to a \code{SpatialPolygons} object.  The code that creates the data file from the raw data can be found at \code{system.file("raw data/polyMaps/create_AKboundary.R",package = "ZVHdata")}. It was subsequently converted to \code{sf} using \code{st_as_sf} function.
}
\usage{
data(AKboundary)
}
\description{
A boundary file for Alaska, USA, of object type \code{sf} from \code{sf} package.
}
\examples{
library(ZVHdata)
library(sf)
data(AKboundary)
summary(AKboundary)
plot(AKboundary)
}
\keyword{datasets}
