% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haulRoad}
\alias{haulRoad}
\title{A line file for the  haul road to Red Dog Mine, Alaska}
\format{A \code{SpatialLines} object with slots @lines, @bbox, and @proj4string. The projection is in Alaska Albers, EPSG:3338.}
\usage{
data(haulRoad)
}
\description{
A line file for the  haul road to Red Dog Mine, Alaska, USA, with object type \code{SpatialLinesDataFrame} from \code{sp} package.
}
\examples{
data(CAKR)
data(haulRoad)
library(sp)
plot(CAKR)
plot(haulRoad, add = TRUE, lwd = 2, col = 'red')
}
\keyword{datasets}
